+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+* now *PainterController.clearContent()* can accept a clear color. By default it is transparent
+* fixed initial state of PenType. Before it was always [PenType.paintbrush]
+
 ## 1.0.1
 * bug fix: the *controller.getPoints()* ponts list wasn't available in *Painter.onDrawingEnded* callback.
 * added some docs.